Adding Query By Countries
Updated TrackQueryParams to include countries. Added the country arrays that Alex made and appended them to the genresFromArrays map in dont_look_at_me.dart file. Updated getTrackFromSentence method to search for countries when the user selects countries. The countries must take precedence over the genre selections. Searching for mood and country may be an issue when examining countries with less music in our database; for example, Uganda only has one song and it's a happy song. If you're sad and search Uganda you get an error infinite loop in the getTrackFromSentenceMethod.