Fixing Image assets, Like Page, and Music Player overflow
Fixed overflow issues with musicplayer and the like page.  Repaired the add like songs feature so it doesnt add duplicates by ID. Also added an overflow check for long song and artists names. Now it caps the string names of songs and artists at a certain size. Changed all the image assets so they would load in the countries page.